transport: replace bufWriter with bufio.Writer
`transport.bufWriter` is a custom buffered writer implementation used when writing to sockets. The
standard library already provides buffered IO via `bufio.Writer`, so there is an opportunity to
clean up custom code.

`transport.bufWriter` introduced in this pull request:
https://github.com/grpc/grpc-go/pull/1962. Out of the 3 benchmarks cited in that pull request
description, I was only able to run the OSS benchmarks. That pull request claims 20%+ gains in
latency and throughput, but it also introduces a number of other changes to the control buffer, so
it is unclear whether those benefits are introduced by `bufWriter` itself.

This change replaces `transport.bufWriter` with `bufio.Writer` from the standard library. This
simplifies the code. I ran the benchmarks on master (before) and then with this change (after). The
results are marginally better before:

```

```